Fix size of element of rnd is not enough
When we use 64-bit machine, memory offset size is 64-bit.
But maximum of element of rnd is 32-bit.

+++ test-tlb.c
@@ -141,7 +141,7 @@ static void randomize_map(void *map, unsigned long size, unsigned long stride)
 	unsigned int *lastpos, *rnd;
 	int n;
 
-	rnd = calloc(size / stride + 1, sizeof(unsigned int));
+	rnd = calloc(size / stride + 1, sizeof(unsigned long));
 	if (!rnd)
 		die("out of memory");
